Bug corrigido

+++ src/main/java/thelibrary/api/biblioteca/entity/Autor.java
@@ -1,10 +1,14 @@
 package thelibrary.api.biblioteca.entity;
 
 
-import lombok.*;
 import thelibrary.api.biblioteca.dto.autor.DadosCadastroAutor;
+import jakarta.persistence.*;
+import lombok.AllArgsConstructor;
+import lombok.EqualsAndHashCode;
+import lombok.Getter;
+import lombok.NoArgsConstructor;
+import thelibrary.api.biblioteca.entity.enumerable.GeneroLiterario;
 
-import javax.persistence.*;
 import java.time.LocalDate;
 
 @Table(name = "autor")
@@ -16,7 +20,7 @@ import java.time.LocalDate;
 public class Autor {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY) // Indica que o valor do ID será gerado automaticamente.
-    private Long id;
+    private Integer id;
     private String nome;
     private LocalDate dataNascimento;
     private String nacionalidade;
